Use an environment file for configuration
To ease changing the hostname and distro selection, an environment file
'config.env' is added and also an '.env' symbolic link targeting it.

The 'config.env' configuration file is a more explicit way to configure
the environment, as in more recent versions of compose, only an '.env'
file can be used to change environment variables available during the
build and deployment phases, so the need for the '.env' link.